Fix delete: replace broken swap-with-last with tombstones
The previous delete left dangling neighbor links (stale/duplicate
results), could heap-overflow when swapping nodes with mismatched
upper-level allocations, mishandled the entry point (segfault repro:
delete 600 of 3000 then search), and saved links referencing
out-of-range slots.

New design: O(1) tombstone delete via an id->slot open-addressing map.
Dead nodes keep carrying graph connectivity (traversal passes through,
inserts may still link to them) but are filtered from query results,
exact scans, and counts. vecdb_compact() rebuilds without tombstones.
Persistence bumped to v2 (dead bitmap + link-range validation on load;
v1 files still load). Duplicate live ids are now rejected.

Adds tests.py (correctness, persistence, edge cases, 3-cycle churn)
and restores the comparison benchmarks under benchmarks/. Churn: 5
cycles of 30% turnover on 10k vectors held recall@10 at 0.95-0.97.

+++ bench.c
@@ -20,7 +20,7 @@ int main(int argc, char **argv) {
         for (size_t d = 0; d < dim; d++) {
             vec[d] = (float)rand() / RAND_MAX;
         }
-        vecdb_add(db, (int)i, vec);
+        vecdb_add(db, (uint64_t)i, vec);
         free(vec);
     }
 
@@ -31,7 +31,7 @@ int main(int argc, char **argv) {
     }
     VecResult out[10];
     vecdb_search_hnsw(db, q, 10, 100, out);
-    printf("Top result id: %llu distance: %f\n", out[0].id, out[0].dist);
+    printf("Top result id: %llu distance: %f\n", (unsigned long long)out[0].id, out[0].dist);
     free(q);
     vecdb_save(db, "index.vecdb");
     vecdb_free(db);
